day02/spotipy/requirements.txt
spotipy
python-dotenv
ipdb

